requirements.txt
aiohttp
configparser
future

# Basic modules
gensim

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster

pipmaster
pydantic
python-dotenv

setuptools
tenacity

# LLM packages
tiktoken
xlsxwriter>=3.1.0
google-genai

sentence_transformers
tqdm 